pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "suitune"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "Django>=5.0",
    "djangorestframework>=3.15",
    "django-allauth>=0.61",
    "watchdog>=4.0.0",
    "mutagen>=1.47.0"
]
